requirements.txt
flask <= 3.1.0
psycopg2-binary <= 2.9.1
flask-cors <= 3.1.1
google-auth-oauthlib <= 1.2.1
google-api-python-client <= 2.154.0
google-auth <= 2.3.0
google-auth-httplib2 <= 0.2.0